Update Shamptake.cpp
Got rid of the sleep temporarily, don't think we will add it back, and slow mode works. If you are testing, know that where the note intakes changes where it shoots, but is very consistent otherwise. THIS HAS BEEN DETERMINED NOT A PROGRAMMING ISSUE, MECHIES ARE FIXING IT WITH LEXAN SO DON'T EVEN TRY.