Got Arm brake working
Programmed arm brake (the worm box one), it has some mechanical things but works for what we need it right now. NOTE: The brake should be set to locked before a match, as it will immediately attempt to disengage. There currently is no way to engage it again, but it has been programmed and just needs to be called.